+++ packages/mongodb-dialects/spring-criteria/src/main/kotlin/com/mongodb/jbplugin/dialects/springcriteria/SpringCriteriaDialectParser.kt
@@ -11,6 +11,7 @@ import com.mongodb.jbplugin.dialects.springcriteria.QueryTargetCollectionExtract
 import com.mongodb.jbplugin.dialects.springcriteria.QueryTargetCollectionExtractor.extractCollectionFromStringTypeParameter
 import com.mongodb.jbplugin.dialects.springcriteria.QueryTargetCollectionExtractor.or
 import com.mongodb.jbplugin.dialects.springcriteria.aggregationstageparsers.AddFieldsStageParser
+import com.mongodb.jbplugin.dialects.springcriteria.aggregationstageparsers.GroupStageParser
 import com.mongodb.jbplugin.dialects.springcriteria.aggregationstageparsers.MatchStageParser
 import com.mongodb.jbplugin.dialects.springcriteria.aggregationstageparsers.ProjectStageParser
 import com.mongodb.jbplugin.dialects.springcriteria.aggregationstageparsers.SortStageParser
@@ -36,6 +37,7 @@ object SpringCriteriaDialectParser : DialectParser<PsiElement> {
         UnwindStageParser(),
         SortStageParser(),
         AddFieldsStageParser(),
+        GroupStageParser(),
     )
 
     override fun isCandidateForQuery(source: PsiElement) =