+++ packages/mongodb-mql-model/src/main/kotlin/com/mongodb/jbplugin/mql/parser/components/HasFieldReference.kt
@@ -19,8 +19,22 @@ inline fun <reified T : HasFieldReference.FieldReference<S>, S> fieldReference()
         }
     }
 }
+inline fun <reified T : HasFieldReference.FieldReference<S>, S> fieldReferences(): Parser<Node<S>, NoFieldReference, List<T>> {
+    return { input ->
+        val refs = input.components<HasFieldReference<S>>().filter {
+            it.reference is T
+        }.map { it.reference as T }
+
+        if (refs.isNotEmpty()) {
+            Either.right(refs)
+        } else {
+            Either.left(NoFieldReference)
+        }
+    }
+}
 
 fun <S> schemaFieldReference() = fieldReference<HasFieldReference.FromSchema<S>, S>()
+fun <S> schemaFieldReferences() = fieldReferences<HasFieldReference.FromSchema<S>, S>()
 
 private fun <S> gatherSchemaFieldReferenceNodes(node: Node<S>): List<Node<S>> {
     val isSchemaFieldReference = node.component<HasFieldReference<S>>()?.reference is HasFieldReference.FromSchema<S>